feat:stats 초기화시점 = > awake
 초기화시점 = > awake

+++ Client/Assets/Scripts/UI/UIController.cs
@@ -42,7 +42,7 @@ public class UIController: MonoBehaviour
 
     public  void HandlerHp()
     {
-        HpBar.value = Mathf.Lerp(HpBar.value, CurHp / MaxHp, Time.deltaTime * 10);
+        HpBar.value = Mathf.Lerp(HpBar.value, (float)CurHp / (float)MaxHp, Time.deltaTime * 10);
     }
 
     public void InitHpbar(DataContainer dataContainer)
